fix&refactor: search log 삭제시 보여지는 배열에서도 삭제,  SearchLogRenderer 파일 분리
